updated snaps import and disconnect func
